+++ src/main/java/be/uantwerpen/sc/tools/IPathplanning.java
@@ -1,8 +1,11 @@
 package be.uantwerpen.sc.tools;
 
-import be.uantwerpen.sc.models.map.Map;
+import be.uantwerpen.rc.models.map.Map;
+import be.uantwerpen.sc.models.map.MapJson;
 import org.springframework.stereotype.Component;
-
+import org.springframework.stereotype.Service;
+import be.uantwerpen.rc.tools.Vertex;
+import javax.websocket.server.ServerEndpoint;
 import java.util.List;
 
 /**
@@ -12,7 +15,7 @@ import java.util.List;
 @Component
 public interface IPathplanning {
 
-    List<Vertex> Calculatepath(Map map, int start, int stop); //implementhed in pathplanning server
+    List<Vertex> Calculatepath(Map map, long start, long stop); //implementhed in pathplanning server
 
 
 }
